Weekly Assignment 12: Quote Machine Pt.2 Electric Bugaloo
I decided to extend by quote machine by having the use input a category of quote they are interested in recieving. This input is used in the obtaining of the quote via the API. It will then display the quote to the user, the user can then ask the interface to show you the author if they wish. It can also obtain a new quote in the same category or a different category if you so choose. Listed below are all the options for categories, if you do not enter one of these keywords no quote will display.